fixed type

+++ src/module/agents/dgn_agent.py
@@ -105,4 +105,4 @@ class DGNAgent(nn.Module):
     def _get_conv_output_shape(self):  # ignore padding
         h = (self.args.obs_height-self.args.kernel_size)/self.args.stride + 1
         w = (self.args.obs_width-self.args.kernel_size)/self.args.stride + 1
-        return h*w*self.args.conv_out_dim
+        return int(h*w*self.args.conv_out_dim)
